+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- iOS: TabView navigation with ConnectionCoordinator extraction and coordinator caching
+
 ## [0.35.0] - 2026-04-25
 
 ### Added